requirements.txt
# Core API packages
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0

# Official AI SDKs (Better quality & features)
openai==1.3.5
replicate==0.22.0

# Image Processing & Enhancement
Pillow==10.1.0
opencv-python-headless==4.8.1.78
numpy==1.26.2

# Quality Enhancement Tools
scikit-image==0.22.0
